requirements.txt
python-telegram-bot==20.8
requests==2.32.3
schedule==1.2.1
requests-toolbelt==1.0.0 